[REFACTOR] Test 리팩토링
