+++ smeem-api/src/test/java/com/smeem/api/controller/TestControllerTest.java
@@ -2,10 +2,7 @@ package com.smeem.api.controller;
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.smeem.api.common.ApiResponseUtil;
-import com.smeem.api.test.TestController;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import com.smeem.api.test.controller.TestController;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
